# Rework view-data migration for efficiency