--- a/resources/migrations/permissions/view_data.sql
+++ b/resources/migrations/permissions/view_data.sql
@@ -12,113 +12,100 @@ INSERT INTO
     perm_value
   )
 SELECT
-  pg.id AS group_id,
+  dp.group_id,
   'perms/view-data' AS perm_type,
-  mt.db_id AS db_id,
-  mt.schema AS schema_name,
-  mt.id AS table_id,
+  dp.db_id,
+  dp.schema_name,
+  dp.table_id,
   CASE
-    WHEN EXISTS (
-      -- If the table has `no-self-service`, convert it to `legacy-no-self-service`
+    WHEN dp.perm_value = 'no-self-service'
+    AND EXISTS (
       SELECT
         1
       FROM
-        data_permissions dp
+        permissions_group_membership pgm
+        JOIN (
+          SELECT
+            group_id,
+            db_id,
+            CAST(NULL AS INTEGER) AS table_id
+          FROM
+            connection_impersonations
+          UNION
+          SELECT
+            group_id,
+            db_id,
+            CAST(NULL AS INTEGER) AS table_id
+          FROM
+            data_permissions
+          WHERE
+            perm_value = 'block'
+            AND table_id IS NULL
+          UNION
+          SELECT
+            group_id,
+            CAST(NULL AS INTEGER),
+            table_id
+          FROM
+            sandboxes
+        ) AS sp ON pgm.group_id = sp.group_id
       WHERE
-        dp.group_id = pg.id
-        AND dp.db_id = mt.db_id
-        AND dp.table_id = mt.id
-        AND dp.perm_type = 'perms/data-access'
-        AND dp.perm_value = 'no-self-service'
+        pgm.group_id <> dp.group_id
+        AND (
+          sp.db_id IS NULL
+          OR sp.db_id = dp.db_id
+        )
+        AND (
+          sp.table_id IS NULL
+          OR sp.table_id = dp.table_id
+        )
     ) THEN 'legacy-no-self-service'
-    -- Otherwise set the table to `unrestricted`. Sandboxed tables are
-    -- `unrestricted` in `data_permissions`and have the sandbox definition
-    -- stored separately
     ELSE 'unrestricted'
   END AS perm_value
 FROM
-  permissions_group pg
-  CROSS JOIN metabase_table mt
+  data_permissions dp
 WHERE
-  pg.name != 'Administrators'
-  -- Only select group/table combinations where table-level `data-access` perms
-  -- are set AND there is at least one `no-self-service` table in the DB that
-  -- cannot be automatically migrated to `unrestricted`
-  AND EXISTS (
+  dp.perm_type = 'perms/data-access'
+  AND dp.table_id IS NOT NULL;
+
+
+-- If all tables in a DB have the same view-data permission, insert a DB-level view-data permission instead
+WITH
+  ConsistentPermissions AS (
     SELECT
-      1
+      group_id,
+      db_id,
+      MIN(perm_value) AS common_value
     FROM
-      data_permissions dp
+      data_permissions
     WHERE
-      dp.group_id = pg.id
-      AND dp.db_id = mt.db_id
-      AND dp.table_id IS NOT NULL
-      AND dp.perm_type = 'perms/data-access'
-      AND dp.perm_value = 'no-self-service'
+      perm_type = 'perms/view-data'
+      AND table_id IS NOT NULL
+    GROUP BY
+      group_id,
+      db_id
+    HAVING
+      COUNT(DISTINCT perm_value) = 1
   )
-  AND EXISTS (
-    SELECT
-      1
-    FROM
-      permissions_group_membership pgm
-    WHERE
-      pgm.group_id <> pg.id
-      AND pgm.user_id IN (
-        SELECT
-          user_id
-        FROM
-          permissions_group_membership pgm_inner
-        WHERE
-          pgm_inner.group_id = pg.id
-      )
-    AND (
-      EXISTS (
-        -- User in another group with block perms for the DB
-        SELECT
-          1
-        FROM
-          data_permissions dp
-        WHERE
-          dp.group_id = pgm.group_id
-          AND dp.db_id = mt.db_id
-          AND dp.table_id IS NULL
-          AND dp.perm_value = 'block'
-      )
-      OR EXISTS (
-        -- User in another group with impersonation for the DB
-        SELECT
-          1
-        FROM
-          connection_impersonations ci
-        WHERE
-          ci.group_id = pgm.group_id
-          AND ci.db_id = mt.db_id
-      )
-      OR EXISTS (
-        -- User in another group with sandboxing for any table in the DB
-        -- that has `no-self-service` perms
-        SELECT
-          1
-        FROM
-          sandboxes s
-          JOIN metabase_table mt_inner ON mt_inner.id = s.table_id
-        WHERE
-          s.group_id = pgm.group_id
-          AND mt_inner.db_id = mt.db_id
-          AND s.table_id IN (
-            SELECT
-              table_id
-            FROM
-              data_permissions dp_inner
-            WHERE
-              dp_inner.group_id = pg.id
-              AND dp_inner.perm_type = 'perms/data-access'
-              AND dp_inner.perm_value = 'no-self-service'
-              AND dp_inner.table_id IS NOT NULL
-          )
-       )
+INSERT INTO
+  data_permissions (
+    group_id,
+    perm_type,
+    db_id,
+    schema_name,
+    table_id,
+    perm_value
   )
-);
+SELECT
+  cp.group_id,
+  'perms/view-data',
+  cp.db_id,
+  NULL,
+  NULL,
+  cp.common_value
+FROM
+  ConsistentPermissions cp;
+
 
 -- Insert DB-level view data permissions for all groups & DBs that don't have
 -- table-level permissions set.
@@ -222,3 +209,19 @@ WHERE
       AND dp.db_id = md.id
       AND dp.perm_type = 'perms/view-data'
   );
+
+
+-- Remove table-level view-data permissions for groups that have DB-level permissions
+DELETE FROM data_permissions
+WHERE
+  (group_id, db_id) IN (
+    SELECT
+      group_id,
+      db_id
+    FROM
+      data_permissions
+    WHERE
+      table_id IS NULL
+      AND perm_type = 'perms/view-data'
+  )
+  AND table_id IS NOT NULL;