--- a/resources/migrations/permissions/view_data.sql
+++ b/resources/migrations/permissions/view_data.sql
@@ -224,4 +224,5 @@ WHERE
       table_id IS NULL
       AND perm_type = 'perms/view-data'
   )
+AND perm_type = 'perms/view-data'
 AND table_id IS NOT NULL;