This PR restructures the SQL migration which populates the new `view-data` permission in order to avoid a hot spot which was causing stats to timeout when deploying the migration.

Specifically, when we're detecting whether to set `legacy-no-self-service` for a particular group & table combination, we need to check these conditions:
* If any user in the group is also in another group with:
 * `block` data permissions for the DB
 * a connection impersonation policy for the DB 
 * a sandboxing policy for the table

Previously, this was happening in the `WHERE` clause of a cross-join between the `permissions_group` and `metabase_table`, which turned out to be very inefficient. The cross join itself wasn't a huge problem, which is why earlier testing didn't expose this, but adding these conditions slows it down significantly.

I've restructured this script into several distinct operations:
1. First, we insert a `view-data` permission row for every existing `data-access` row that was set on an individual table. To check whether `legacy-no-self-service` should be set, we join `data_permissions` against a subquery that `UNION`s the relevant columns from `data_permissions`, `sandboxes` and `connection_impersonations`. This is a much faster process.
2. Then, we look for any tables in a DB which are set to all the same values and insert a DB-level row with that value.
3. Then we set DB-level `view-data` permissions for any DBs that previously had DB-level `data-access` permissions. This is unchanged from the previous implementation because it wasn't seeing the same performance issues.
4. Finally, we delete any table-level rows which also have DB-level rows set. (This is necessary as clean-up from step 2)

I've also had to split the `view_data` migration into two versions, one for postgres & h2 and one for mysql & mariadb due to syntax compatibility issues.

**Timing data**
This data is collected from running both the `view_data` and the (unchanged) `create_queries` migrations on a local Postgres app DB populated with a dump of data from stats. tl;dr: I've cut the runtime down from 2 minutes to 3 seconds.

Previous migration:
```
metabase=# \i resources/migrations/permissions/view_data.sql
INSERT 0 0
Time: 129350.615 ms (02:09.351)
INSERT 0 435
Time: 32.590 ms
metabase=# \i resources/migrations/permissions/create_queries.sql
INSERT 0 424
Time: 21.807 ms
INSERT 0 16796
Time: 227.459 ms

```

New migration: 
```
metabase=# \i resources/migrations/permissions/view_data.sql
INSERT 0 16796
Time: 3844.136 ms (00:03.844)
INSERT 0 11
Time: 6.521 ms
INSERT 0 424
Time: 175.322 ms
DELETE 16796
Time: 22.087 ms
metabase=# \i resources/migrations/permissions/create_queries.sql
INSERT 0 424
Time: 21.439 ms
INSERT 0 16796
Time: 218.129 ms

```